move prompts to ui layer
